--- a/pkg/ofac/mapper.go
+++ b/pkg/ofac/mapper.go
@@ -233,7 +233,7 @@ func parseAddresses(inputs []Address) []search.Address {
 		// {... Address:"place du Lac 2", CityStateProvincePostalCode:"Geneve 1204", Country:"Switzerland", AddressRemarks:""}
 		var addr search.Address
 		addr.Line1 = inputs[i].Address
-		addr.Country = inputs[i].Country
+		addr.Country = country.Normalize(inputs[i].Country)
 
 		parts := strings.Fields(inputs[i].CityStateProvincePostalCode)
 		if len(parts) == 1 {