Each .Country field in a query can be normalized before search and each index's .Country can be normalized in the mapper.

<details>
<summary>Benchmarks</summary>

```
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz

                                │ ./pkg/search/before.txt │        ./pkg/search/after.txt        │
                                │         sec/op          │    sec/op     vs base                │
API_Search/normal-16                         23.65m ± ∞ ¹   23.36m ± ∞ ¹       ~ (p=1.000 n=1) ²
API_Search/debug-16                          65.77m ± ∞ ¹   50.99m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/1-16                83.87m ± ∞ ¹   72.03m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/3-16                33.34m ± ∞ ¹   31.47m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/5-16                25.91m ± ∞ ¹   24.06m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/10-16               22.25m ± ∞ ¹   20.33m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/20-16               17.93m ± ∞ ¹   16.28m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/25-16               18.20m ± ∞ ¹   16.65m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/50-16               16.76m ± ∞ ¹   15.94m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/100-16              15.28m ± ∞ ¹   14.91m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/150-16              15.17m ± ∞ ¹   16.50m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/200-16              15.32m ± ∞ ¹   16.98m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/250-16              19.42m ± ∞ ¹   17.49m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/dynamic_group_size-16                21.40m ± ∞ ¹   18.66m ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                      23.93m         22.35m        -6.60%

                                │ ./pkg/search/before.txt │        ./pkg/search/after.txt         │
                                │          B/op           │     B/op       vs base                │
API_Search/normal-16                        8.602Mi ± ∞ ¹   8.589Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
API_Search/debug-16                         48.47Mi ± ∞ ¹   51.86Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/1-16               7.136Mi ± ∞ ¹   7.136Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/3-16               7.136Mi ± ∞ ¹   7.136Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/5-16               7.136Mi ± ∞ ¹   7.136Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/10-16              7.137Mi ± ∞ ¹   7.137Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/20-16              7.138Mi ± ∞ ¹   7.138Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/25-16              7.138Mi ± ∞ ¹   7.138Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/50-16              7.142Mi ± ∞ ¹   7.141Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/100-16             7.146Mi ± ∞ ¹   7.146Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/150-16             7.150Mi ± ∞ ¹   7.150Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/200-16             7.155Mi ± ∞ ¹   7.155Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/250-16             7.160Mi ± ∞ ¹   7.159Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/dynamic_group_size-16               7.138Mi ± ∞ ¹   7.138Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                     8.299Mi         8.338Mi        +0.47%

                                │ ./pkg/search/before.txt │        ./pkg/search/after.txt        │
                                │        allocs/op        │  allocs/op    vs base                │
API_Search/normal-16                         447.8k ± ∞ ¹   447.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
API_Search/debug-16                          659.4k ± ∞ ¹   660.2k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/1-16                365.4k ± ∞ ¹   365.4k ± ∞ ¹       ~ (p=1.000 n=1) ³
_Search/fixed_group_size/3-16                365.4k ± ∞ ¹   365.4k ± ∞ ¹       ~ (p=1.000 n=1) ³
_Search/fixed_group_size/5-16                365.4k ± ∞ ¹   365.4k ± ∞ ¹       ~ (p=1.000 n=1) ³
_Search/fixed_group_size/10-16               365.4k ± ∞ ¹   365.4k ± ∞ ¹       ~ (p=1.000 n=1) ³
_Search/fixed_group_size/20-16               365.5k ± ∞ ¹   365.5k ± ∞ ¹       ~ (p=1.000 n=1) ³
_Search/fixed_group_size/25-16               365.5k ± ∞ ¹   365.5k ± ∞ ¹       ~ (p=1.000 n=1) ³
_Search/fixed_group_size/50-16               365.5k ± ∞ ¹   365.5k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/100-16              365.6k ± ∞ ¹   365.6k ± ∞ ¹       ~ (p=1.000 n=1) ³
_Search/fixed_group_size/150-16              365.7k ± ∞ ¹   365.7k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/200-16              365.8k ± ∞ ¹   365.8k ± ∞ ¹       ~ (p=1.000 n=1) ³
_Search/fixed_group_size/250-16              365.9k ± ∞ ¹   365.9k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/dynamic_group_size-16                365.5k ± ∞ ¹   365.5k ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                      386.9k         386.9k        +0.01%

pkg: github.com/moov-io/watchman/pkg/search
                                     │ ./pkg/search/before.txt │        ./pkg/search/after.txt         │
                                     │         sec/op          │    sec/op     vs base                 │
DebugSimilarity/individuals-16                    29.96µ ± ∞ ¹   10.32µ ± ∞ ¹        ~ (p=1.000 n=1) ²
DebugSimilarity/individuals-debug-16              56.22µ ± ∞ ¹   34.91µ ± ∞ ¹        ~ (p=1.000 n=1) ²
DebugSimilarity/businesses-16                    14.050µ ± ∞ ¹   6.668µ ± ∞ ¹        ~ (p=1.000 n=1) ²
DebugSimilarity/businesses-debug-16               36.11µ ± ∞ ¹   26.06µ ± ∞ ¹        ~ (p=1.000 n=1) ²
DebugSimilarity/vessels-16                        2.736µ ± ∞ ¹   2.544µ ± ∞ ¹        ~ (p=1.000 n=1) ²
DebugSimilarity/vessels-debug-16                  17.40µ ± ∞ ¹   16.40µ ± ∞ ¹        ~ (p=1.000 n=1) ²
DebugSimilarity/aircraft-16                      121.86µ ± ∞ ¹   39.69µ ± ∞ ¹        ~ (p=1.000 n=1) ²
DebugSimilarity/aircraft-debug-16                203.77µ ± ∞ ¹   96.31µ ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                           31.66µ         17.78µ        -43.85%

                                     │ ./pkg/search/before.txt │        ./pkg/search/after.txt         │
                                     │          B/op           │     B/op       vs base                │
DebugSimilarity/individuals-16                   1.734Ki ± ∞ ¹   1.688Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/individuals-debug-16             8.329Ki ± ∞ ¹   8.281Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/businesses-16                    1.199Ki ± ∞ ¹   1.172Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/businesses-debug-16              7.551Ki ± ∞ ¹   7.520Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/vessels-16                         184.0 ± ∞ ¹     184.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
DebugSimilarity/vessels-debug-16                 6.335Ki ± ∞ ¹   6.334Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/aircraft-16                      6.969Ki ± ∞ ¹   6.734Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/aircraft-debug-16                20.57Ki ± ∞ ¹   20.32Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                          3.477Ki         3.430Ki        -1.33%

                                     │ ./pkg/search/before.txt │       ./pkg/search/after.txt        │
                                     │        allocs/op        │  allocs/op   vs base                │
DebugSimilarity/individuals-16                     135.0 ± ∞ ¹   129.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/individuals-debug-16               190.0 ± ∞ ¹   184.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/businesses-16                      72.00 ± ∞ ¹   70.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/businesses-debug-16                107.0 ± ∞ ¹   105.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/vessels-16                         16.00 ± ∞ ¹   16.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
DebugSimilarity/vessels-debug-16                   34.00 ± ∞ ¹   34.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
DebugSimilarity/aircraft-16                        496.0 ± ∞ ¹   476.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
DebugSimilarity/aircraft-debug-16                  631.0 ± ∞ ¹   611.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                            116.4         113.5        -2.44%
```

</details>

![Screenshot 2025-02-14 at 13 38 16](https://github.com/user-attachments/assets/141f6255-5c41-434a-aab2-22706233b259)
![Screenshot 2025-02-14 at 13 38 41](https://github.com/user-attachments/assets/9ef20889-a559-4db7-bdb9-990e49220935)